Update README.md
E - represents energy (measured in Joules), m - represents mass (measured in kilograms), and  c - represents the speed of light (measured approximately as 300000000 meters per second), per Albert Einstein et al. 

Essentially, the formula means that mass and energy are equivalent.

Python that prompts the user for mass as an integer (in kilograms) and then outputs the equivalent number of Joules as an integer.  Assume that the user will input an integer.